Improved install.

+++ install.sh
@@ -11,6 +11,21 @@ fi
 if [ ! -d $HOME/.gnome2/gedit/snippets ]
 then
     mkdir -p ~/.gnome2/gedit/snippets
+    cp $FOLDER/snippets/* ~/.gnome2/gedit/snippets/
+else
+    if [ ! -e ~/.gnome2/gedit/snippets/python.xml ]
+    then
+        cp $FOLDER/snippets/* ~/.gnome2/gedit/snippets/
+    else
+        echo -n 'The file python.xml already exist, do you want to replace it (y/n)? '
+        read option
+        if [ $option = 'y' ]
+        then
+            cp $FOLDER/snippets/* ~/.gnome2/gedit/snippets/
+            echo 'File replaced.'
+        else
+            echo 'File not replaced.'
+        fi
+    fi
 fi
-cp $FOLDER/snippets/* ~/.gnome2/gedit/snippets/
